feat(statistics): Improve statistics data

+++ src/main/java/com/ticketgo/dto/RevenueStatisticsDTO.java
@@ -3,6 +3,8 @@ package com.ticketgo.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -11,6 +13,6 @@ import lombok.*;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class RevenueStatisticsDTO {
     private String period;
-    private Double totalRevenue;
+    private BigDecimal totalRevenue;
     private Long totalTicketsSold;
 }
